+++ test/numeric/num_test.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_test/flutter_test.dart';
 import 'package:formdator/formdator.dart';
+import 'package:test/test.dart';
 
 void main() {
   const error = 'The value must be numeric';
@@ -200,8 +200,8 @@ void main() {
       });
 
       test('illegal range, min >= max', () {
-        expect(() => Num.range(10, 1), throwsAssertionError);
-        expect(() => Num.range(10, 10), throwsAssertionError);
+        expect(() => Num.range(10, 1), throwsA(isA<AssertionError>()));
+        expect(() => Num.range(10, 10), throwsA(isA<AssertionError>()));
       });
       test('valid input', () {
         expect(range('5'), null);